# client/public/swagger.yaml
openapi: 3.0.0
info:
  description: Api documentation
  version: 1.0.1
  title: aim-container API
...
servers:
  - url: http://127.0.0.1:43801/{basePath}
    variables:
      basePath:
        default: api/v1
    description: Local env
security:
  - tokenAuth:
paths:
...
  "/tags/update":
    post:
      operationId: updateTag
      summary: Update tag's name or color.
      tags:
        - Tag
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                id:
                  type: string
                name:
                  type: string
                color:
                  type: string
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                type: object
...
  "/commits/tags/{commit_id}":
    get:
      operationId: getCommitTags
      summary: Get commit tags
      tags:
        - Commit
      parameters:
          - name: commit_id
            in: path
            required: true
            schema:
              type: string
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                type: object
  "/commits/tags/update":
    post:
      operationId: updateCommitTag
      summary: Update commit tag
      tags:
        - Commit
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                commit_hash:
                  type: string
                experiment_name:
                  type: string
                tag_id:
                  type: string
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                type: object
  "/commits/info/{experiment}/{commit_id}":
    get:
      operationId: getCommitInfo
      summary: Get commit info
      tags:
        - Commit
      parameters:
        - name: commit_id
          in: path
          required: true
          schema:
            type: string
        - name: experiment
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                type: object
  "/commits/search?q={query}":
    get:
      operationId: getCommitsByQuery
      summary: Search commits
      tags:
        - Commit
      parameters:
          - name: query
            in: path
            required: false
            schema:
              type: string
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                type: object
...
components:
  securitySchemes:
    tokenAuth:
      type: apiKey
      name: Authorization
      in: header
...